Fix different currencies and donation workflow
The paypal donate buttons might support different currencies out of the box,
but not without configuring it somehow.
As this was already planned to be moved to a solution which works without
client side scripts and opening a new window, this commit changes that already.
This also fixes using different currencies (which was supported before by the
backend already)

Fixes #36