Do not fail updates to database
As noted in the recent releaes, starting the donation tool after
updating it, an error message like:
(node:13568) UnhandledPromiseRejectionWarning: Error: alter table `<table-name>` [...] `<column-name>` [...] - SQLITE_ERROR: duplicate column name: [...]
(Use `node --trace-warnings ...` to show where the warning was created)
(node:13568) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). To terminate the node process on unhandled promise rejection, use the CLI flag `--unhandled-rejections=strict` (see https://nodejs.org/api/cli.html#cli_unhandled_rejections_mode). (rejection id: 2)

could occur. This happened when a database migration was needed and the
tool automatically attempted to do so. The error did not indicate a data
integrity problem or alike, but was annoying and not very trustworthy
(real errors could be overseen because of that).

This commit fixes the underlying issue of this error and should therefor
prevent it from happening again.